Fix inorrect typecasts

+++ questions/Vectors/server.py
@@ -19,8 +19,8 @@ def generate(data):
 
     # add them to data["params"]
     data["params"]["theta"] = theta
-    data["params"]["r"] = int(r[1])
-    data["params"]["s"] = int(s[0])
+    data["params"]["r"] = float(r[1])
+    data["params"]["s"] = float(s[0])
     data["params"]["rscale"] = rscale
     data["params"]["sscale"] = sscale
 
@@ -30,8 +30,8 @@ def generate(data):
     result = rot_matrix.dot(rscale * r + sscale * s)
 
     # add the correct answers for the x and y components to data["correct-answers"]
-    data["correct_answers"]["x-comp"] = int(result[0])
-    data["correct_answers"]["y-comp"] = int(result[1])
+    data["correct_answers"]["x-comp"] = float(result[0])
+    data["correct_answers"]["y-comp"] = float(result[1])
 
 def file(data):
     if data["filename"] == "figure.png":
